WIP autoindexing

+++ inc/Utils.hpp
@@ -3,7 +3,7 @@
 #include <sys/stat.h>
 #include <vector>
 
-std::string              toString(int n);
+std::string              toString(long long n);
 void                     replace(std::string& str, const std::string& from, const std::string& to);
 std::vector<std::string> split(const std::string& s, char delimiter);
 bool                     startsWith(const std::string& str, const std::string& prefix);
